% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_rental_listings.R
\name{get_rental_listings}
\alias{get_rental_listings}
\title{Get Zillow rental listings time series data}
\usage{
get_rental_listings(bedrooms = 1, type = NULL, geography = "Zip",
  rate = "Total")
}
\arguments{
\item{bedrooms}{a numeric value specifying the number of bedrooms. If not needed, leave at the default (1)}

\item{type}{(optional) a character string specifying housing type, from NULL,'SFR','Multi','Duplex','Condo/Co-op','Studio', and 'SFR/Condo'.}

\item{geography}{string specifying the desired geographic region to summarise. Choices are 'Metro','City','State','Neighborhood','Zip', and 'County'.}

\item{rate}{a string specifying the rate, either 'Total' or 'PerSqFt'}
}
\value{
A tibble
}
\description{
reads the static .csv file for the desired Zillow Home Values series based on data type and geography, by building a url path to the .csv file hosted by Zillow
}
\examples{

get_rental_listings(bedrooms=5, rate='PerSqFt', type="Studio",geography="Zip")

get_rental_listings(bedrooms=1, rate='Total', type="Multi", geography="State")
}
