% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDeepComps.R
\name{GetDeepComps}
\alias{GetDeepComps}
\title{Get comparable recent property sales for a given property with additional property data}
\usage{
GetDeepComps(zpid, count = 10, rentzestimate = FALSE, api_key,
  raw = FALSE)
}
\arguments{
\item{zpid}{The Zillow property id to search.}

\item{count}{(integer) How many comparables to return?}

\item{rentzestimate}{(logical) If \code{TRUE}, gets the rent zestimate.}

\item{api_key}{A character string specifying your unique Zillow API key}

\item{raw}{(logical) If \code{TRUE} the raw XML data from the API call is returned (i.e., the original ZillowR call)}
}
\value{
If \code{raw=T}, a raw XML document. If \code{raw=F} (default), a data frame with columns corresponding to address information, Zestimates, and property information. The number of columns varies by property use type.
}
\description{
For a given Zillow property ID, find valuation data for comparable recent property sales,
and return both the property valuation estimate as well as other property characteristics for each property
}
\examples{

zpid='1340244'

\dontrun{

GetDeepComps(zpid, count=10, rentzestimate=TRUE,
 api_key= getOption('ZillowR-zws_id'), raw=FALSE)

  }

}
