% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDeepSearchResults.R
\name{GetDeepSearchResults}
\alias{GetDeepSearchResults}
\title{Get the 'Deep Search' results from the Zillow API}
\usage{
GetDeepSearchResults(address, city = NULL, state = NULL,
  zipcode = NULL, rentzestimate = F, api_key, raw = F)
}
\arguments{
\item{address}{(required) A character string specifying the address}

\item{city}{An optional character string specifying the city name}

\item{state}{An optional character string specifying the state name (abbreviated length 2)}

\item{zipcode}{An optional character or numeric string of length 5 specifying the zip code}

\item{rentzestimate}{if \code{TRUE}, gets the rent zestimate.}

\item{api_key}{(required) A unique character string specifying the Zillow user's API key}

\item{raw}{(optional) If \code{TRUE}, the raw xml document from the API call is returned. If \code{FALSE}, the data is extracted and formatted into a data frame.}
}
\value{
If \code{raw=T}, the XML document directly from the API call. If \code{raw=F} (default), a data frame with columns corresponding to address information, Zestimates, and property information. The number of columns varies by property use type.
}
\description{
For a given address, extract property information including building data and Zestimates.
At least one of zipcode or city/state information must be included.
}
\examples{
\dontrun{zapi_key = getOption('ZillowR-zws_id')}

\dontrun{

GetDeepSearchResults(address='600 S. Quail Ct.', zipcode='67114',
 rentzestimate=FALSE, api_key=zapi_key)

 }

\dontrun{GetDeepSearchResults(address='312 Hayward Ave.', city='Ames', state='IA',
 rentzestimate=TRUE, zipcode='50014', api_key=zapi_key) }

}
