% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDeepSearchResults_dataframe.R
\name{GetDeepSearchResults_dataframe}
\alias{GetDeepSearchResults_dataframe}
\title{Get the 'Deep Search' results from the Zillow API for a dataframe of addresses}
\usage{
GetDeepSearchResults_dataframe(.df, col.address, col.zipcode = NULL,
  col.city = NULL, col.state = NULL, rentzestimate = F, api_key,
  raw = FALSE)
}
\arguments{
\item{.df}{a dataframe with address data}

\item{col.address}{a numeric value specifying which column in \code{.df} contains the addresses}

\item{col.zipcode}{a numeric value specifying which column contains the zipcodes (required if no city/state information)}

\item{col.city}{a numeric value specifying which column contains the city (required if no zipcode information)}

\item{col.state}{a numeric value specifying which column contains the state abbrev. (required if no zipcode information)}

\item{rentzestimate}{if \code{TRUE}, additional zestimates for rental prices are returned.}

\item{api_key}{a character string specifying the unique Zillow API key}

\item{raw}{logical, if \code{TRUE}, a list with raw XML documents  (i.e., the original ZillowR call) for each property in the data frame is returned}
}
\value{
A data frame with columns corresponding to address, property, and Zestimate information
}
\description{
A wrapper for \code{GetDeepSearchResults} that takes data frames with address information as inputs rather than a single address,
and returns the search results for all addresses
}
\examples{
\dontrun{
zapi_key = getOption('ZillowR-zws_id')

adrs <- c('733 Normandy Ct', '600 South Quail Ct','2902 Wood St', '3425 Locust St.')

locations <- data.frame(address=adrs,zipcode=c('67114','67114','50014', '64109'),
state=c("KS","KS","IA","MO"), city=c("Newton","Newton","Ames","Kansas City"))

GetDeepSearchResults_dataframe(locations, col.address=1, col.zipcode=2,
col.city=4, col.state=3,api_key=zapi_key)

}
}
