% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetUpdatedPropertyDetails.R
\name{GetUpdatedPropertyDetails}
\alias{GetUpdatedPropertyDetails}
\title{Make request to Zillow API GetUpdatedPropertyDetails Web Service}
\usage{
GetUpdatedPropertyDetails(zpid, api_key, raw = FALSE)
}
\arguments{
\item{zpid}{The Zillow Property ID for the property for which to obtain
information. Required.}

\item{api_key}{character string specifying Zillow API key}

\item{raw}{logical, if \code{TRUE} the raw XML data from the API call is returned (i.e., the original ZillowR call)}
}
\value{
The result set contains the following attributes:

\itemize{
  \item Property address
  \item Zillow property identifier
  \item Posting details such as the agent name, MLS number, price, and
      posting type (For Sale or Make Me Move(tm))
  \item Up to five photos of the property
  \item Updated home facts such as beds, baths, square footage, home
      description, and neighborhood and school names
}

A data frame with columns corresponding to zpid, Date, and updated property information
}
\description{
For a specified property, the GetUpdatedPropertyDetails API returns all of
the home facts that have been edited by the home's owner or agent.
}
\examples{
\dontrun{
zapi_key = getOption('ZillowR-zws_id')
zpid='48749425'

GetUpdatedPropertyDetails(zpid,api_key= zapi_key)
}
}
