% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetZestimates.R
\name{GetZestimates}
\alias{GetZestimates}
\alias{GetZestimate}
\title{Get Zestimates for a given Zillow property ID}
\usage{
GetZestimate(zpids, rentzestimate = FALSE, api_key, raw = FALSE)
}
\arguments{
\item{zpids}{a single value or vector of Zillow property ids}

\item{rentzestimate}{if \code{TRUE}, gets the rent zestimate.}

\item{api_key}{character string specifying Zillow API key}

\item{raw}{logical, if \code{TRUE} the raw XML data from the API call is returned (i.e., the original ZillowR call)}
}
\value{
A data frame with columns corresponding to zpid, Date, and Zestimate information
}
\description{
A wrapper for \code{GetDeepSearchResults} that takes data frames with address information as inputs rather than a single address,
and returns the search results for all addresses
}
\examples{
library(tidyverse)

zapi_key = 'X1-ZWz181ckl6u9e3_9k5bc'

zpid='1341571'


GetZestimate(zpids=zpid,rentzestimate=TRUE,api_key=zapi_key)
}
