% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newtonhouses}
\alias{newtonhouses}
\title{Data on 51 comparable homes in the Newton, KS, USA area}
\format{A data frame with 51 rows and 26 variables:
\describe{
\item{address}{Property street address}
\item{zipcode}{Five digit zipcode}
\item{city}{City}
\item{state}{State abbreviation}
\item{lat}{Latitude of location}
\item{long}{Longitude of location}
\item{region_name}{}
\item{region_id}{Zillow-assigned five digit region ID}
\item{type}{Type of region (i.e., Neighborhood, City,...)}
\item{zestimate}{Zillow's estimate of the home value}
\item{zest_lastupdated}{The date of the last update to the Zestimate}
\item{zest_monthlychange}{Zestimate change from previous month}
\item{zest_percentile}{Percentile rank of the property Zestimate within the region}
\item{zestimate_low}{The lower bound of the range of valuation}
\item{zestimate_high}{The upper bound of the range of valuation}
\item{zpid}{The unique Zillow-assigned property ID number}
\item{bathrooms}{Number of bathrooms in the property}
\item{bedrooms}{Number of bedrooms in the property}
\item{finishedSqFt}{Size of finished area in the property, in square feet}
\item{lastSoldDate}{Date the property was last sold (if applicable)}
\item{lastSoldPrice}{The most recent sale price of the property (if applicable)}
\item{lotSizeSqFt}{Size of the lot the property is on, in square feet}
\item{taxAssessment}{The most recent available local property tax assessment value, in dollars}
\item{taxAssessmentYear}{Year corresponding to the most recent property tax assessment}
\item{totalRooms}{Total number of rooms in the property}
\item{yearBuilt}{Year the property was built}
  ...
}}
\source{
\url{http://www.zillow.com/}
}
\usage{
newtonhouses
}
\description{
A dataset containing property attributes and value information for 210 homes in the Newton, KS area. This dataset was generated by chaining together the `GetDeepSearchResults` and `GetDeepComps` functions (as shown in the package vignette) to find comparable homes starting from the initial single address of 600 S. Quail Ct.
}
\keyword{datasets}
