% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetZestimates.R
\name{GetZestimates}
\alias{GetZestimates}
\alias{GetZestimate}
\title{Get Zestimates for a given Zillow property ID}
\usage{
GetZestimate(zpids, rentzestimate = FALSE, api_key, raw = FALSE)
}
\arguments{
\item{zpids}{a single value or vector of Zillow property IDs}

\item{rentzestimate}{if \code{TRUE}, gets the rent zestimate.}

\item{api_key}{character string specifying Zillow API key}

\item{raw}{logical, if \code{TRUE} the raw XML data from the API call is returned (i.e., the original ZillowR call)}
}
\value{
A data frame with columns corresponding to zpid, Date, and Zestimate information
}
\description{
Returns a dataframe with the Zillow zestimate and/or rental valuation information for a given Zillow property ID.
}
\examples{
\dontrun{
zapi_key = getOption('ZillowR-zws_id')

zpid='1341571'

GetZestimate(zpids=zpid,rentzestimate=TRUE,api_key=zapi_key)
}
}
